requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
Django==4.1.3
elastic-transport==8.4.0
elasticsearch==8.5.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fake-useragent==1.1.1
fastapi==0.85.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
gensim==4.3.1
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.29
googletransx==2.4.2
h11==0.13.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==5.1.0
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
multitasking==0.0.11
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
Plost==0.2.5
plotly==5.14.1
protobuf==3.20.3
pyarrow==10.0.1
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-socks==2.0.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
schedule==1.1.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snscrape==0.4.3.20220106
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.20.4
streamlit==1.21.0
streamlit-lottie==0.0.3
summarization==0.0.1
tenacity==8.2.2
textblob==0.17.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.13.0
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
uvicorn==0.18.3
validators==0.20.0
watchdog==2.2.0
webdriver-manager==3.8.5
wordcloud==1.8.2.2
wsproto==1.2.0
yarl==1.8.2
yfinance==0.1.87
zipp==3.11.0
